LayoutTests/fast/forms/date/calendar-picker-appearance-pre-100.html

<!DOCTYPE html>
<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<input type=date id=date value="0002-02-02">
<script>
function sendKey(input, keyName) {
    var event = document.createEvent('KeyboardEvent');
    event.initKeyboardEvent('keydown', true, true, document.defaultView, keyName);
    input.dispatchEvent(event);
}

function finishTest() {
    pickerWindow.removeEventListener('resize', finishTest);
    sendKey(pickerWindow.global.daysTable._daysContainer, 'Right');
    sendKey(pickerWindow.global.daysTable._daysContainer, 'Enter');
    shouldBe('document.getElementById("date").value', '"0002-02-03"');
    setTimeout(function() { finishJSTest(); }, 0);
}

description('Check if year before 100 can be set.');

if (window.testRunner)
    testRunner.waitUntilDone();

window.jsTestIsAsync = true;

if (window.internals)
    internals.settings.setEnableMockPagePopup(true);

var dateInput = document.getElementById('date');
dateInput.offsetTop;
sendKey(dateInput, 'Down');
var pickerWindow = document.getElementById('mock-page-popup').contentWindow;
pickerWindow.addEventListener('resize', finishTest, false);
var successfullyParsed = true;
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
